% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lean2.R
\name{as.data.frame.lean2}
\alias{as.data.frame.lean2}
\title{Convert lean2 object into a data.frame}
\usage{
\method{as.data.frame}{lean2}(x, ...)
}
\arguments{
\item{x}{An object of class \code{lean2}}

\item{\dots}{Passed down to \code{as.data.frame()}}
}
\description{
Convert lean2 object into a data.frame
}
\details{
Function \code{as.lean2()} and \code{as.data.frame.lean2()} perform
inverse operations. However, the results will only be identical
if there are no duplicates, if all items are integer, and if
all auxiliary variables are numeric.
}
\seealso{
\code{\link{as.lean2}}, \code{\link{as.data.frame}}
}
