% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lean2.R
\name{as.lean2}
\alias{as.lean2}
\alias{read_lean2}
\title{Convert a data frame with one row per visit into a long list}
\usage{
as.lean2(
  x,
  visit_id = c("subjid", "agedays"),
  remove_duplicates = FALSE,
  strict = TRUE
)

read_lean2(dbfile)
}
\arguments{
\item{x}{A `data.frame` with one visit per row with
administrative, covariates, items and auxiliary variables.}

\item{visit_id}{A character vector with the names of the grouping variable(s)
that define a visit. The default is \code{c("subjid", "agedays")}.}

\item{remove_duplicates}{A logical indicating whether the function
may proceed silently by removing rows with duplicate keys.}

\item{strict}{A logical specifying whether the resulting item
names must conform to one of the built-in names in the `dscore`
package. The default is \code{strict = TRUE}.}

\item{dbfile}{File name of the database}
}
\description{
The `as.lean2()` function converts a data set with one row
per visit into a list with the following elements:

The `read_lean2()` function construct an object of class `lean2` from
a database.
}
\details{
1. `visits`: a data frame with administrative and covariate variables. The
minimum columns are specified by `visit_id` (e.g., `"subjid"`, `"agedays"`).
Each visit is defined by a unique combination of these columns.
2. `responses`: a data frame with items in long format. The minimal
columns are specified by `visit_id` plus columns named `"item"` and
`"response"` that stores each item-response score.
3. `auxiliary`: a data frame with auxiliary variables in long format,
4. `visit_id`: a character vector with the names of the grouping variable(s)
whose unique combination define a visit.

The object has class `lean2`. This function is a modern version of the
`gseddata::as.lean()` function.
}
\note{
Function \code{as.lean()} and \code{as.data.frame.lean()} perform
inverse operations. However, the results will only be identical
if there are no duplicate visits, if all responses are integer, if
all auxiliary variables are numeric, and if no item in \code{x} has
entirely missing data.
}
\examples{
data <- data.frame(subjid = c(1L, 1L, 2L, 3L, 3L),
                   agedays = c(100L, 200L, 149L, 120L, 200L),
                   sexn = c(2L, 2L, 1L, 1L, 1L),
                   by3cgd036 = c(0L, 1L, NA, NA, 0L),
                   by3cgd037 = c(1L, NA, NA, NA, 1L),
                   by3cgd038 = c(NA, 1L, 0L, NA, 0L),
                   outcome1 = c(NA, NA, 121, 113, 113),
                   outcome2 = c(2, 3, NA, 8, 6))
lean <- as.lean2(data)
lean
unlean <- as.data.frame(lean)
unlean
identical(data, unlean)
}
\seealso{
\code{\link{as.data.frame.lean2}}
}
\author{
Stef van Buuren
}
