% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_administrative.R
\name{calculate_administrative}
\alias{calculate_administrative}
\title{Calculate administrative variables}
\usage{
calculate_administrative(data)
}
\arguments{
\item{data}{Long form data, typically created by scripts/edit_data.R}
}
\value{
Five administrtative variables
}
\description{
Calculate administrative variables
}
