% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_DIF_table.R
\name{calculate_DIF_table}
\alias{calculate_DIF_table}
\title{Calculate DIF statistic per item for multiple groups}
\usage{
calculate_DIF_table(data, items, score, group, levels = NULL, thr = NULL)
}
\arguments{
\item{data}{Data frame with dichotomous item responses and age}

\item{items}{Character vector with item names}

\item{score}{Data frame with item parameters}

\item{group}{Factor or character vector with group names. Must have same
length as the number of rows in the data frame.}

\item{levels}{Optional. Character vector with group names. If omitted, the function
will use the unique values of the group variable.}

\item{thr}{Optional. Threshold for DIF classification using the MH-statistics.
If omitted, the function will use the threshold from the `difGMH()` function.}
}
\value{
data frame with DIF statistics per item
}
\description{
This function calculates DIF statistics for multiple groups using the
`difR` package. The function uses the `difGenLogistic()` function for
generalized logistic regression and the `difGMH()` function for the
generalized Mantel-Haenszel test. The function returns a table with
DIF statistics per item.
}
\details{
The function uses the `"holm"` method for p-value adjustment.
}
\examples{
library(dscore)
data <- dscore::milestones
score <- dscore(data = data, xname = "age", metric = "logit")[["d"]]
items <- intersect(get_itemnames(ins = "ddi", order = "indm"), colnames(data))

calculate_DIF_table(data = data, items = items, group = data$sex, score = score)
}
