% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_wide.R
\name{make_wide}
\alias{make_wide}
\title{Merge measurement made on different days to one child level record}
\usage{
make_wide(
  long,
  instruments = c("gpa", "gto", "by3"),
  days = 10L,
  drop_na_age = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{long}{Long form data, typically created by scripts/edit_data.R}

\item{instruments}{Character vector of instruments names. Currently supported
instruments are `"gpa"`, `"gto"` and `"by3"`.}

\item{days}{Non-negative integer, window width in number of days}

\item{drop_na_age}{Logical. Should measurements with missing ages be dropped?}

\item{quiet}{Logical. Print messages to terminal?}
}
\value{
A data frame with one record per child. The first five columns
are administrative variables, followed by the intervals in days per
instrument. The remaining columns are items from the instruments.
}
\description{
For a given child, this script joins measurements made within a
specified window.
}
\examples{
\dontrun{
joined <- make_wide(long, instruments = c("gpa", "gto"))
}
}
