% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require_package.R
\name{require_package}
\alias{require_package}
\title{Require a package}
\usage{
require_package(pkg, version, repo = "CRAN")
}
\arguments{
\item{pkg}{A string with the required package}

\item{version}{A string with the minimum version, e.g. "0.1.0"}

\item{repo}{A string with the GitHub repo name. For the default
`repo == "CRAN"` installation proceeds from CRAN.}
}
\value{
A logical indicating success (TRUE) or failure (FALSE) for version check.
}
\description{
Check whether a package is installed. If not, the
procedure offers to install the latest version from CRAN or GitHub.
Then it check for a minimum version.
}
