% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_DIF_table.R
\name{calculate_DIF_table}
\alias{calculate_DIF_table}
\title{Calculate DIF statistic per item for multiple groups}
\usage{
calculate_DIF_table(
  data,
  items,
  group,
  score = "score",
  levels = NULL,
  thr = NULL
)
}
\arguments{
\item{data}{Data frame with dichotomous item responses and age}

\item{items}{Character vector with item names}

\item{group}{Factor or character vector with group names. Must have same
length as the number of rows in the data frame.}

\item{score}{Data frame with item parameters. Optional.}

\item{levels}{Optional. Character vector with group names. If omitted, the function
will use the unique values of the group variable.}

\item{thr}{Optional. Threshold for DIF classification using the MH-statistics.
If omitted, the function will use the threshold from the `difGMH()` function.}
}
\value{
data frame with DIF statistics per item
}
\description{
This function calculates DIF statistics for multiple groups using the
`difR` package. The function uses the `difGenLogistic()` function for
generalized logistic regression and the `difGMH()` function for the
generalized Mantel-Haenszel test. The function returns a table with
DIF statistics per item.
}
\details{
The function uses the `"holm"` method for p-value adjustment.
}
\examples{
\dontrun{
# For some reason, the following example does not work with the
# version 5.1 of difR. The example is kept here for future reference.
# We get: Error in n.ppk * diag(rk) - rk \%*\% t(rk) : non-conformable arrays

# DIF by sex, Dutch SMOCC data
library(dscore)
library(childdevdata)
data <- childdevdata::gcdg_nld_smocc
items <- intersect(get_itemnames(ins = "ddi", order = "indm"), colnames(data))
score <- dscore(data = data, items = items, xname = "agedays", xunit = "days",
metric = "logit")[["d"]]
group <- as.character(data$sex)

# Remove rows with fewer than 3 observations
# Remove missing values from group and score
responses <- as.matrix(data[, items])
sparse <- rowSums(!is.na(responses)) < 3
idx <- !(is.na(score) | is.na(group) | sparse)
responses <- responses[idx, ]
group <- group[idx]
score <- score[idx]

library(difR )
LR <- difLogistic(Data = responses,
  group = group,
  focal.names = unique(group),
  match = score)
MH <- difMH(Data = responses,
  group = group,
  focal.names = unique(group),
  match = score)
}
}
