% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_required_packages.R
\name{update_required_packages}
\alias{update_required_packages}
\title{Update the required packages to their development versions}
\usage{
update_required_packages()
}
\description{
This function checks if the required packages are installed and up-to-date.
If not, it asks the user if they want to install the development versions
from GitHub.
}
